# server/resource/Cleverbuys_db/dealership.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE dealership.js PLEASE EDIT ../custom/dealershipCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_Cleverbuys_db = require('../../db/Cleverbuys_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/dealershipCustom.js');

/*
 * SCHEMA DB dealership
 * 
	{
		email: {
			type: 'String', 
			required : true
		},
		name: {
			type: 'String', 
			required : true
		},
		telephone: {
			type: 'Number'
		},
		town: {
			type: 'String', 
			required : true
		},
		website: {
			type: 'String'
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		brands: {
			type: Schema.ObjectId,
			ref : "dealership"
		},
		
	}
 * 
 */



//CRUD METHODS


/**
 * dealershipService.create
 *   @description CRUD ACTION create
 *   @param dealership obj Object to insert
 *
 */
app.post(properties.api + '/dealership', function (req, res) {
	obj = new db_Cleverbuys_db.dealership(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * dealershipService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id dealership
 *
 */
app['delete'](properties.api + '/dealership/:id', function (req, res) {
	db_Cleverbuys_db.dealership.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * dealershipService.findBybrands
 *   @description CRUD ACTION findBybrands
 *   @param Objectid key Id of model to search for
 *
 */
app.get(properties.api + '/dealership/findBybrands/:key', function (req, res) {

	db_Cleverbuys_db.dealership.find({ 'brands' : req.params.key}).exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
	
});

/**
 * dealershipService.get
 *   @description CRUD ACTION get
 *   @param ObjectId id Id dealership
 *   @returns dealership
 *
 */
app.get(properties.api + '/dealership/:id', function (req, res) {
	db_Cleverbuys_db.dealership.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * dealershipService.list
 *   @description CRUD ACTION list
 *   @returns ARRAY OF dealership
 *
 */
app.get(properties.api + '/dealership', function (req, res) {
	db_Cleverbuys_db.dealership.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * dealershipService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id dealership
 *   @returns dealership
 *
 */
app.post(properties.api + '/dealership/:id', function (req, res) {
	db_Cleverbuys_db.dealership.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */