# server/resource/Cleverbuys_db/state.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE state.js PLEASE EDIT ../custom/stateCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_Cleverbuys_db = require('../../db/Cleverbuys_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/stateCustom.js');

/*
 * SCHEMA DB state
 * 
	{
		name: {
			type: 'String', 
			required : true
		},
		shortName: {
			type: 'String', 
			required : true
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		
	}
 * 
 */



//CRUD METHODS


/**
 * stateService.create
 *   @description CRUD ACTION create
 *   @param state obj Object to insert
 *
 */
app.post(properties.api + '/state', function (req, res) {
	obj = new db_Cleverbuys_db.state(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * stateService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id state
 *
 */
app['delete'](properties.api + '/state/:id', function (req, res) {
	db_Cleverbuys_db.state.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * stateService.get
 *   @description CRUD ACTION get
 *   @param ObjectId id Id state
 *   @returns state
 *
 */
app.get(properties.api + '/state/:id', function (req, res) {
	db_Cleverbuys_db.state.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * stateService.list
 *   @description CRUD ACTION list
 *   @returns ARRAY OF state
 *
 */
app.get(properties.api + '/state', function (req, res) {
	db_Cleverbuys_db.state.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * stateService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id state
 *   @returns state
 *
 */
app.post(properties.api + '/state/:id', function (req, res) {
	db_Cleverbuys_db.state.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */