# server/swagger.yaml
openapi: 3.0.0
info:
  title: Cleverbuys
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  User  -----
...
  /dealership/:
    post:
      summary: CRUD ACTION create
      tags:
        - dealership
      parameters:
        - name: obj
          in: path
          description: Object to insert
          required: true
          schema:
            type: dealership
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dealership"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                telephone:
                  type: number
                town:
                  type: string
                website:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - dealership
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/dealership"
  /dealership/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - dealership
      parameters:
        - name: id
          in: path
          description: Id dealership
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - dealership
      parameters:
        - name: id
          in: path
          description: Id dealership
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dealership"
    post:
      summary: CRUD ACTION update
      tags:
        - dealership
      parameters:
        - name: id
          in: path
          description: Id dealership
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dealership"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dealership"
  /dealership/findBybrands/{key}:
    get:
      summary: CRUD ACTION findBybrands
      tags:
        - dealership
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/dealership"
    
  #-----  state  -----
  /state/:
    post:
      summary: CRUD ACTION create
      tags:
        - state
      parameters:
        - name: obj
          in: path
          description: Object to insert
          required: true
          schema:
            type: state
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/state"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                shortName:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - state
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/state"
  /state/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - state
      parameters:
        - name: id
          in: path
          description: Id state
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - state
      parameters:
        - name: id
          in: path
          description: Id state
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/state"
    post:
      summary: CRUD ACTION update
      tags:
        - state
      parameters:
        - name: id
          in: path
          description: Id state
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/state"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/state"
    
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    User:
      properties:
        _id:
          type: string
        lat:
          type: number
        lng:
          type: number
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        postCode:
          type: string
        roles:
          type: string
        state:
          type: string
        surname:
          type: string
        town:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username
...
    dealership:
      properties:
        _id:
          type: string
        email:
          type: string
        name:
          type: string
        telephone:
          type: number
        town:
          type: string
        website:
          type: string
      required:
        -  _id
        - email
        - name
        - town
    state:
      properties:
        _id:
          type: string
        name:
          type: string
        shortName:
          type: string
      required:
        -  _id
        - name
        - shortName